use the needed size for each variable

+++ HomeworkLesson17/PrintAgeWithSum/PrintAgeWithSum.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int age = 25;
+    unsigned short age = 25;
     std::cout << "After 5 years you will be " << age + 5 << " years old.\n";
 
     return 0;
